fix:changed type names and refactring Modal Func
